refactor: container를 이용한 리팩토링 작업 완료
