+++ repeat/step2/src/clients/index.ts
@@ -1,1 +1,1 @@
-export * from "./youtubeRestClient";
+export * from "./YoutubeRestClient";
