+++ repeat/step2/src/repositories/LectureVideoRepository.ts
@@ -1,10 +1,10 @@
 import {Repository} from "~@core";
 import {LectureVideo} from "~@domain";
+import {Injectable} from "~@core/container";
 
+@Injectable
 export class LectureVideoRepository extends Repository<LectureVideo[]> {
   constructor() {
     super('LectureVideo');
   }
 }
-
-export const lectureVideoRepository = new LectureVideoRepository();
