+++ repeat/step2/src/repositories/RecentSearchesRepository.ts
@@ -1,9 +1,9 @@
 import {Repository} from "~@core";
+import {Injectable} from "~@core/container";
 
+@Injectable
 export class RecentSearchRepository extends Repository<string[]> {
   constructor() {
     super('RecentSearches');
   }
 }
-
-export const recentSearchesRepository = new RecentSearchRepository();
